chore(metadata.config.ts): add keywords, site, and images to metadata configuration
The keywords, site, and images properties were added to the metadata configuration in order to improve the SEO and social media sharing capabilities of the application. The keywords property now includes relevant keywords related to the festival, such as "festival", "musique", "concert", and others. The site property was added to specify the name of the website. The images property was added to include the path to the logo image that will be used for social media sharing.