feat(sponsors.config.ts): add sponsors configuration file
A new file `sponsors.config.ts` has been added to the `src/config` directory. This file contains the configuration for sponsors. It exports a `Sponsor` type and an array of `sponsors` objects. Each `sponsor` object has properties such as `name`, `imgPath`, and `href` which represent the name, image path, and hyperlink for a sponsor respectively.

This configuration file is necessary to store and manage the information of sponsors for the application.