feat: add billetterie.config.ts file to store billetterie link
The billetterie.config.ts file is added to the config directory. This file exports a constant `billetterieLink` which contains the URL "https://cuicuitedays.fr". This file is created to store the billetterie link in a centralized location for easy access and maintenance.

fix: update billetterie link in footer, navbar, and home section

The billetterie link in the footer, navbar, and home section is updated to use the `billetterieLink` constant from the `billetterie.config.ts` file. This ensures that the correct billetterie link is used consistently across the application. The link is updated to "https://cuicuitedays.fr" to match the value in the `billetterie.config.ts` file.